1) domain changed to save txt for every time step 2) temporal equation changed to delete garbage pointer 3) domain changed to save only this and last time index grid 4) double to float 5) cn rect solver inherits fe rect solver and erase redundant parts

+++ src/initial_condition/initial_condition.h
@@ -1,16 +1,16 @@
 #pragma once
 //#include "src/domain/base_domain.h"
-#include "src/domain/base_domain_copy.h"
+#include "src/domain/rect_domain.h"
 #include <functional>
 #include <complex>
 #include <iostream>
 class InitialCondition
 {
     private:
-        std::function<std::complex<double>(double, double)> initial_condition_function;
+        std::function<std::complex<float>(float, float)> initial_condition_function;
     public:
         InitialCondition() = default;
-        InitialCondition(std::function<std::complex<double>(double, double)> initial_condition_function);
+        InitialCondition(std::function<std::complex<float>(float, float)> initial_condition_function);
         void assign_to_domain(BaseDomain *domain);
         
 };
