+++ src/test/test_psolver/test_fe_rect_psolver.cpp
@@ -8,19 +8,19 @@
 bool test_fe_rect_psolver(int rank, int size){
     std::cout<<"[Processor "<<rank<<"] Start"<<std::endl;
     bool all_passed = true;
-    std::function<double(double, double)> potential;
+    std::function<float(float, float)> potential;
 
-    double g;
+    float g;
     RectangularDomain* domain = (new RectangularDomain(101, 101, 0, 0.1, 101, -5, 5, -5, 5));
-    auto initial_cond_function = [](double x, double y)
-    { return std::complex<double>{1*std::exp(-(x*x + y*y)/(9))}; };
+    auto initial_cond_function = [](float x, float y)
+    { return std::complex<float>{1*std::exp(-(x*x + y*y)/(9))}; };
 
     auto *initial_condition =new  InitialCondition(initial_cond_function);
 
     initial_condition-> assign_to_domain(domain);
     
-    potential= [](double x, double y ){
-        return (double)  (1. *x*x + 2. * y *y);  };
+    potential= [](float x, float y ){
+        return (float)  (1. *x*x + 2. * y *y);  };
     g = 1. ;
 
     //std::cout << "." << std::endl;
@@ -48,8 +48,8 @@ bool test_fe_rect_psolver(int rank, int size){
 
     solver.solve(std::to_string(rank));
     //At first step, psi(0,0) = 1+ i (4 e^-1 - 5) 
-    double real = 1. ; 
-    double imag = 4 * std::exp(-1./9.) - 5; 
+    float real = 1. ; 
+    float imag = 4 * std::exp(-1./9.) - 5; 
     if (!is_close((*domain).at(10, 10, 1)->wave_function.real(), real, 1e-12))
     {
         all_passed = false;
