+++ src/test/test_solver/test_cn_rect_solver.cpp
@@ -10,16 +10,16 @@ bool test_cn_solver_creation()
 
     RectangularDomain *domain = (new RectangularDomain(41, 41, 0, 1, 101, -5, 5, -5, 5));
 
-    auto initial_cond_function = [](double x, double y)
-    { return std::complex<double>{1 * std::exp(-(x * x + y * y) / (9))}; };
+    auto initial_cond_function = [](float x, float y)
+    { return std::complex<float>{1 * std::exp(-(x * x + y * y) / (9))}; };
     auto *initial_condition = new InitialCondition(initial_cond_function);
     initial_condition->assign_to_domain(domain);
 
-    auto potential = [](double x, double y)
+    auto potential = [](float x, float y)
     {
-        return (double)0.5 * (x * x + y * y);
+        return (float)0.5 * (x * x + y * y);
     };
-    double g = -1;
+    float g = -1;
     CNRectSolver solver = CNRectSolver(potential, g, domain);
 
     solver.solve(1e-13, 101);
