+++ src/utils.cpp
@@ -1,5 +1,5 @@
 #include "utils.h"
 
-bool is_close(double a, double b, double tolerance){
+bool is_close(float a, float b, float tolerance){
     return std::abs(a - b) < tolerance;
 }
